Converted hold to double

+++ Recursion.java
@@ -8,7 +8,7 @@ public class Recursion{
 
     */
     public static double sqrt(double n, double tolerance){
-      if(n>=0){int hold = n;return sqrt(-1*n/2,tolerance);}
+      if(n>=0){double hold = n;return sqrt(-1*n/2,tolerance);}
       n*=-1;
       if(Math.abs(n*n-hold)/hold<=tolerance)return n;
       return sqrt((hold/n+n)/2,tolerance);
